+++ src/Interfaces/Strings/AlfStringRead.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Alf\Interfaces\Strings;
+
+use Alf\Manipulator\AlfStringManipulator;
+
+interface AlfStringRead extends AlfStringGet {
+
+    public function refManipulator() : AlfStringManipulator;
+
+}