+++ src/Interfaces/Strings/AlfStringWork.php
@@ -2,6 +2,10 @@
 
 namespace Alf\Interfaces\Strings;
 
-interface AlfStringWork extends AlfStringGet, AlfStringSet {
+interface AlfStringWork extends AlfStringRead, AlfStringSet {
+
+    public function getStringLength() : int;
+
+    public function getStringByteSize() : int;
 
 }